+++ StabilityMatrix.Avalonia/ViewModels/PackageManager/PackageInstallDetailViewModel.cs
@@ -14,6 +14,7 @@ using Microsoft.Extensions.Logging;
 using StabilityMatrix.Avalonia.Controls;
 using StabilityMatrix.Avalonia.Extensions;
 using StabilityMatrix.Avalonia.Languages;
+using StabilityMatrix.Avalonia.Models.PackageSteps;
 using StabilityMatrix.Avalonia.Services;
 using StabilityMatrix.Avalonia.ViewModels.Base;
 using StabilityMatrix.Core.Attributes;
@@ -188,6 +189,7 @@ public partial class PackageInstallDetailViewModel(
         }
 
         var prereqStep = new SetupPrerequisitesStep(prerequisiteHelper, pyRunner, SelectedPackage);
+        var unpackSiteCustomizeStep = new UnpackSiteCustomizeStep(Path.Combine(installLocation, "venv"));
 
         var downloadOptions = new DownloadPackageVersionOptions();
         var installedVersion = new InstalledPackageVersion();
@@ -252,6 +254,7 @@ public partial class PackageInstallDetailViewModel(
             setPackageInstallingStep,
             prereqStep,
             downloadStep,
+            unpackSiteCustomizeStep,
             installStep,
             setupModelFoldersStep,
             addInstalledPackageStep